Refactor Process::Error in C++ wrapper.
We use a class (instead of enum) with static instances available from
inside the wrapper. The main advantage is that we are able to assign
the static values in the cpp file to the values of PROCESS_LIB_ERROR.
This means we can move members of PROCESS_LIB_ERROR around without
worrying about breaking Process::Error. We can't assign the values
in process.hpp because this would mean including process.h which
makes all the other C api declarations available when including
process.hpp which we want to avoid.