Rework POSIX `wait_timeout` to use `sigtimedwait`.
`sigtimedwait` is used to wait on `SIGCHLD` signals with a timeout.
After each `SIGCHLD` signal, we can check if the process we're waiting
for has exited. If it hasn't exited, we call `sigtimedwait` again until
the timeout expires.

This is more efficient than spawning a new process just to wait
on an existing process with a timeout.

We use kqueue on OSX as it does not support `sigtimedwait`.

Because of this change, we also stop putting each process into its own
process group since we're not waiting on multiple processes anymore
using `waitpid`.