Always search for executables relative to parent working directory.
Until now, if a working directory different from the parent directory
was specified, on Windows reproc would search for executable specified
as a relative path relative to the parent process working directory but
on Linux reproc would search relative to the child process working
directory. This commit makes sure relative executable paths are always
searched for relative to the parent process working directory. This is
accomplished by prepending the parent process working directory to the
executable's path if it is relative and a custom working directory is
specified before calling `chdir`.