Update `process::start` to work with any string type.
Instead of excepting the underlying type to be `std::string`,
`process::start` now works with any string type that exposes a `size`
and index operator. This comes at a small performance cost due to the
extra allocations required but we think this is a good tradeoff.

`working_directory` is now a `const char *` instead of a  `std::string`
pointer again to work with any string type that can be converted to a
null-terminated C string.

This commit also cleans up the SFINAE in `process::start`.