Pass actual array as initial value in parse functions.
When passing NULL values, users had to special case for this when using
`memcpy` in sinks or parsers as it does not accept `NULL`. By passing an
actual address with a zero size, users do not have to special case for
`NULL`.

The drain example has been updated to not special case for `NULL`
anymore.