POSIX: Duplicate standard streams in `redirect` (#26)
Child fd's returned by `redirect` are closed in the parent process after they have been duplicated to the child process. Since we return the standard stream fd's directly from `redirect` this means that we're closing the parent process's streams which we don't want to do. Instead of checking whether `redirect` was used to inherit the parent's streams, we duplicate the standard stream fd's in `redirect` to allow child fd's returned by `redirect` to be closed after they've been inherited by the child process regardless of the redirection type used.