reproc: Refactor `process_wait` to work with multiple processes.
`process_wait` now waits until the first process in a given list of
processes exits. This is preparation work for adding a new function to
reproc that waits for the first process of a list of processes to exit.

The Windows implementation implements the new API but only handles a
single process for now.