reproc(++): Remove `reproc_parse/process::parse`.
Its behaviour can be easily replicated with
`reproc_drain/process::drain`. If a protocol is being parsed, the
current message of the parser will not be complete which indicates an
error or an unlimited amount of data is being read in which case there's
no reason to treat `REPROC_ERROR_STREAM_CLOSED/error::stream_closed`
differently.