reproc: Remove the code for waiting for more than one process.
The implementations of waiting for multiple processes to exit
concurrently added too many gotchas with relation to multithreading. On
POSIX, until now we relied on signals which are global and as a result
should be handled by the user program instead of a library.

Removing these changes makes it easier to plug reproc into programs for
running one off programs. Should waiting for multiple processes to exit
from a single thread be needed, it is better to use an event loop such
as libuv which has support for spawning child processes.

reproc will aim itself fully at being easy to use in multithreaded
programs without requiring lots of platform specific setup by users or
having annoying limitations such as processes only being useable from
the thread they were created on.