reproc: Fix and simplify io test.
When writing to both stdout and stderr, we wrongly assumed all output
from stdout would be read before all the output from stderr. Now, we
always write to both stdout and stderr and capture the streams in
separate sinks so the comparisons always succeed regardless of the order
data is read from the streams.