reproc: Work around Windows data loss issue.
On Windows, redirecting the stdout/stderr of a child process can
result in data loss because Windows doesn't correctly close the
redirected socket when the child process exits. See
https://developercommunity.visualstudio.com/content/problem/977949/child-process-redirect-stdio-to-sockets-stdio-not.html
for a description of this issue. To solve this, we keep open an
extra handle to the redirected stdout/stderr sockets in the parent
process. When we detect the child process has exited, we close these
extra handles correctly ourselves which keeps all child process
output intact. To detect whether the child process has exited, we
re-use the exit pipe. This pipe can safely be closed by the child
process because no data is actually transferred over this pipe.

The disadvantage of this solution is that we can't detect anymore
when a child process on Windows closes its standard streams but
doesn't exit.