Fix _WIN32_WINNT macro redefinition warnings on MSVC
Windows project can be configured using the _WIN32_WINNT macro. This macro
can be defined using the /D compiler option.  When using the precompiler
definition and the warning level is at least 1, then macro redefinition
warning (c4005) is produced by the compiler.

The ifndef block is able to detect if the macro symbol was already defined
and ensures that the appropriate version of Windows was selected.

https://docs.microsoft.com/en-us/windows/win32/winprog/using-the-windows-headers
https://docs.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-level-1-c4005